Add high contrast colors
Adds high contrast colors, based on the colors used in the editor.
The goal here is mainly to make appkit elements (and especially the
HeaderMenu) usuable with the high contrast mode in the editor.
As such, this commit does not completely solve the high contrast
color issue, and will need futher work in the future.

Of particular note is that high contrast modes like to use their
aggressive "focus" color in a lot of cases where non-high-contrast-modes
prefer more subtle "neutral" colors. This currently results in a lot of
annoying conditionals in the code, which we may want to avoid in the
future.