Improve `match` and remove its `fallback` overload
Unfortunately, `match` had a problem where sometimes it was unsound.
This was not really the fault of `match` as TS has questionable
semantics regarding `Record<string, _>`. The following worked before:

    const foo: number = match("bla", {
        "a": () => 3,
    });

So: no fallback, an incomplete object and still, it returns `number`.
This can of course fail at runtime. The problem is that `Record<X, _>`
works differently for `string` and for unions of strings.

The new version solves this by returning `Out | null` if not all cases
are covered, meaning that for `string` it always returns `Out | null`.

I also removed the `fallback` overload as the same can easily be
achieved via `?? fallback`.

I also added a rudimentary compile-test to make sure the problematic
cases actually cause compile errors.